+++ app/app/alerts/[id]/page.tsx
@@ -1,0 +1,2 @@
+import AlertDetail from "../../../alert-detail";
+export default function AlertDetailPage(){return <AlertDetail/>;}